resources: add allocation and deallocation methods
Documentation states the resource class must provide allocation and
deallocation methods but the API does not honor it.

This has also a side effect: if the resources initialisation fail, the
already allocated resources won't be cleaned up.

Added methods to honour the documentation and fixed resources leakage in
case of initialisation error.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>